Move credentials to new package in core.api
