--- core.api/src/main/java/org/mqnaas/core/api/Credentials.java
@@ -1,5 +1,0 @@
-package org.mqnaas.core.api;
-
-public class Credentials {
-
-}
