+++ core.api/src/main/java/org/mqnaas/core/api/credentials/Credentials.java
@@ -1,0 +1,5 @@
+package org.mqnaas.core.api.credentials;
+
+public class Credentials {
+
+}
