+++ mqnaas-test-helpers/src/main/java/org/mqnaas/test/helpers/clientprovider/TestInternalClientProvider.java
@@ -1,8 +1,8 @@
 package org.mqnaas.test.helpers.clientprovider;
 
 import org.mqnaas.clientprovider.api.client.IInternalClientProvider;
-import org.mqnaas.core.api.Credentials;
 import org.mqnaas.core.api.Endpoint;
+import org.mqnaas.core.api.credentials.Credentials;
 
 /**
  * {@link IInternalClientProvider} for testing purposes able to return always a new {@link EmptyClient}.
